+++ CHANGELOG.md
@@ -168,3 +168,4 @@
 * Remove all `Show` deriving instances (library users can use standalone deriving)
 * Rename `ask'` to `environment` for access user defined read-only environment
 * Remove `with` package from `extra-deps`, put open product to `Utils` module
+* Replace raw `Int` and `Int 64` on `ID Sender` and `ID Chat` in `Kick`, `Unban`, `Restrict` and `Promote` methods